move ScimUser.Name into separate class.
