+++ scim/src/main/java/org/cloudfoundry/identity/uaa/scim/ScimName.java
@@ -1,0 +1,77 @@
+package org.cloudfoundry.identity.uaa.scim;
+
+import org.codehaus.jackson.map.annotate.JsonSerialize;
+
+@JsonSerialize(include = JsonSerialize.Inclusion.NON_NULL)
+public final class ScimName {
+    String formatted;
+
+    String familyName;
+
+    String givenName;
+
+    String middleName;
+
+    String honorificPrefix;
+
+    String honorificSuffix;
+
+    public ScimName() {
+    }
+
+    public ScimName(String givenName, String familyName) {
+        this.givenName = givenName;
+        this.familyName = familyName;
+        this.formatted = givenName + " " + familyName;
+    }
+
+    public String getFormatted() {
+        return formatted;
+    }
+
+    public void setFormatted(String formatted) {
+        this.formatted = formatted;
+    }
+
+    public String getFamilyName() {
+        return familyName;
+    }
+
+    public void setFamilyName(String familyName) {
+        this.familyName = familyName;
+    }
+
+    public String getGivenName() {
+        return givenName;
+    }
+
+    public void setGivenName(String givenName) {
+        this.givenName = givenName;
+    }
+
+    public String getMiddleName() {
+        return middleName;
+    }
+
+    public void setMiddleName(String middleName) {
+        this.middleName = middleName;
+    }
+
+    public String getHonorificPrefix() {
+        return honorificPrefix;
+    }
+
+    public void setHonorificPrefix(String honorificPrefix) {
+        this.honorificPrefix = honorificPrefix;
+    }
+
+    public String getHonorificSuffix() {
+        return honorificSuffix;
+    }
+
+    public void setHonorificSuffix(String honorificSuffix) {
+        this.honorificSuffix = honorificSuffix;
+    }
+
+}
+
