+++ scim/src/main/java/org/cloudfoundry/identity/uaa/scim/ScimUser.java
@@ -39,82 +39,9 @@ import org.springframework.util.StringUtils;
 @JsonDeserialize(using = ScimUserJsonDeserializer.class)
 public final class ScimUser extends ScimCore {
 
-    @JsonSerialize(include = JsonSerialize.Inclusion.NON_NULL)
-    public static final class Name {
-        String formatted;
-
-        String familyName;
-
-        String givenName;
-
-        String middleName;
-
-        String honorificPrefix;
-
-        String honorificSuffix;
-
-        public Name() {
-        }
-
-        public Name(String givenName, String familyName) {
-            this.givenName = givenName;
-            this.familyName = familyName;
-            this.formatted = givenName + " " + familyName;
-        }
-
-        public String getFormatted() {
-            return formatted;
-        }
-
-        public void setFormatted(String formatted) {
-            this.formatted = formatted;
-        }
-
-        public String getFamilyName() {
-            return familyName;
-        }
-
-        public void setFamilyName(String familyName) {
-            this.familyName = familyName;
-        }
-
-        public String getGivenName() {
-            return givenName;
-        }
-
-        public void setGivenName(String givenName) {
-            this.givenName = givenName;
-        }
-
-        public String getMiddleName() {
-            return middleName;
-        }
-
-        public void setMiddleName(String middleName) {
-            this.middleName = middleName;
-        }
-
-        public String getHonorificPrefix() {
-            return honorificPrefix;
-        }
-
-        public void setHonorificPrefix(String honorificPrefix) {
-            this.honorificPrefix = honorificPrefix;
-        }
-
-        public String getHonorificSuffix() {
-            return honorificSuffix;
-        }
-
-        public void setHonorificSuffix(String honorificSuffix) {
-            this.honorificSuffix = honorificSuffix;
-        }
-
-    }
-
     private String userName;
 
-    private Name name;
+    private ScimName name;
 
     private List<ScimEmail> emails;
 
@@ -153,7 +80,7 @@ public final class ScimUser extends ScimCore {
     public ScimUser(String id, String userName, String givenName, String familyName) {
         super(id);
         this.userName = userName;
-        this.name = new Name(givenName, familyName);
+        this.name = new ScimName(givenName, familyName);
     }
 
     public String getUserName() {
@@ -172,11 +99,11 @@ public final class ScimUser extends ScimCore {
         this.userName = userName;
     }
 
-    public Name getName() {
+    public ScimName getName() {
         return name;
     }
 
-    public void setName(Name name) {
+    public void setName(ScimName name) {
         this.name = name;
     }
