+++ uaa/src/test/java/org/cloudfoundry/identity/uaa/integration/LoginServerSecurityIntegrationTests.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- *     Cloud Foundry 
+ *     Cloud Foundry
  *     Copyright (c) [2009-2014] Pivotal Software, Inc. All Rights Reserved.
  *
  *     This product is licensed to you under the Apache License, Version 2.0 (the "License").
@@ -20,6 +20,7 @@ import java.util.Collections;
 import java.util.Map;
 
 import org.cloudfoundry.identity.uaa.message.PasswordChangeRequest;
+import org.cloudfoundry.identity.uaa.scim.ScimName;
 import org.cloudfoundry.identity.uaa.scim.ScimUser;
 import org.cloudfoundry.identity.uaa.test.TestAccountSetup;
 import org.cloudfoundry.identity.uaa.test.UaaTestAccounts;
@@ -48,7 +49,7 @@ import org.springframework.web.client.RestTemplate;
 /**
  * Integration test to verify that the Login Server authentication channel is
  * open and working.
- * 
+ *
  * @author Dave Syer
  */
 public class LoginServerSecurityIntegrationTests {
@@ -96,7 +97,7 @@ public class LoginServerSecurityIntegrationTests {
 
         ScimUser user = new ScimUser();
         user.setUserName(JOE);
-        user.setName(new ScimUser.Name("Joe", "User"));
+        user.setName(new ScimName("Joe", "User"));
         user.addEmail("joe@blah.com");
 
         ResponseEntity<ScimUser> newuser = client.postForEntity(serverRunning.getUrl(userEndpoint), user,
